Fix Safari issue
Thanks for this repo. 

In the Safari browser (Mac and iOS) the SWFs weren't playing, whereas on Chrome they were.

Turns out it was because the load event for the window object had already fired by the time the useEffect hook ran, causing the event listener inside useRuffle to never trigger since the event had already passed.

The modified useRuffle function here checks if the document is already loaded and if so, executes the code, and if not delays and checks again.

It is now working well in Safari.

Thanks again.